+++ java/compaction/compaction-datafusion/pom.xml
@@ -23,7 +23,7 @@
         <version>0.32.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>compaction-rust</artifactId>
+    <artifactId>compaction-datafusion</artifactId>
     <dependencies>
         <dependency>
             <groupId>com.github.jnr</groupId>
