+++ java/compaction/compaction-datafusion/src/main/java/sleeper/compaction/df/DataFusionFunctions.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.rust;
+package sleeper.compaction.df;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import jnr.ffi.Struct;
