+++ java/compaction/compaction-datafusion/src/test/java/sleeper/compaction/df/DataFusionCompactionRunnerLocalStackIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.rust;
+package sleeper.compaction.df;
 
 import org.apache.parquet.hadoop.ParquetReader;
 import org.apache.parquet.hadoop.ParquetWriter;
@@ -25,7 +25,7 @@ import org.testcontainers.containers.localstack.LocalStackContainer;
 import sleeper.compaction.core.job.CompactionJob;
 import sleeper.compaction.core.job.CompactionJobFactory;
 import sleeper.compaction.core.job.CompactionRunner;
-import sleeper.compaction.rust.RustCompactionRunner.AwsConfig;
+import sleeper.compaction.df.DataFusionCompactionRunner.AwsConfig;
 import sleeper.core.partition.PartitionsBuilder;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TableProperties;
@@ -62,7 +62,7 @@ import static sleeper.core.properties.testutils.TablePropertiesTestHelper.create
 import static sleeper.core.schema.SchemaTestHelper.createSchemaWithKey;
 import static sleeper.core.statestore.testutils.StateStoreUpdatesWrapper.update;
 
-public class RustCompactionRunnerLocalStackIT extends LocalStackTestBase {
+public class DataFusionCompactionRunnerLocalStackIT extends LocalStackTestBase {
 
     private final InstanceProperties instanceProperties = createTestInstanceProperties();
     private final TableProperties tableProperties = createTestTablePropertiesWithNoSchema(instanceProperties);
@@ -105,7 +105,7 @@ public class RustCompactionRunnerLocalStackIT extends LocalStackTestBase {
     }
 
     protected RecordsProcessed compact(CompactionJob job) throws Exception {
-        CompactionRunner runner = new RustCompactionRunner(createAwsConfig());
+        CompactionRunner runner = new DataFusionCompactionRunner(createAwsConfig());
         return runner.compact(job, tableProperties, stateStore.getPartition(job.getPartitionId()));
     }
