+++ java/compaction/pom.xml
@@ -14,8 +14,9 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <artifactId>aws</artifactId>
         <groupId>sleeper</groupId>
@@ -31,7 +32,7 @@
         <module>compaction-job-creation</module>
         <module>compaction-job-creation-lambda</module>
         <module>compaction-job-execution</module>
-        <module>compaction-rust</module>
+        <module>compaction-</module>
         <module>compaction-tracker</module>
         <module>compaction-task-creation</module>
     </modules>
