+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/investigate/InvestigateDataFusionCompactionOOM.java
@@ -23,7 +23,6 @@ import software.amazon.awssdk.services.s3.S3Client;
 import sleeper.compaction.core.job.CompactionJob;
 import sleeper.compaction.core.job.CompactionJobSerDe;
 import sleeper.compaction.core.job.CompactionRunner;
-import sleeper.compaction.rust.RustCompactionRunner;
 import sleeper.core.iterator.IteratorCreationException;
 import sleeper.core.partition.PartitionTree;
 
@@ -42,7 +41,7 @@ import java.util.UUID;
  * cd rust
  * cargo build --release --target=x86_64-unknown-linux-gnu
  * cd ../java
- * mvn clean install -am -pl compaction/compaction-rust -Pquick -DskipRust
+ * mvn clean install -am -pl compaction/compaction-datafusion -Pquick -DskipRust
  * </pre>
  */
 public class InvestigateDataFusionCompactionOOM {
@@ -74,7 +73,7 @@ public class InvestigateDataFusionCompactionOOM {
         Path tempDir = Files.createTempDirectory("sleeper-test");
         Path outputFile = tempDir.resolve(UUID.randomUUID().toString());
         CompactionJob localJob = inferredJob.toBuilder().outputFile("file://" + outputFile.toString()).build();
-        CompactionRunner runner = new RustCompactionRunner();
+        CompactionRunner runner = new DataFusionCompactionRunner();
         runner.compact(localJob, logs.tableProperties(), partitionTree.getPartition(localJob.getPartitionId()));
     }
